sparql/src/main/java/io/github/basilapi/basil/sparql/SpecificationFactory.java
/*
 * Copyright (c) 2021. Enrico Daga and Luca Panziera
 *
 * MLicensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
 *  You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 *  Unless required by applicable law or agreed to in writing, software
 *  distributed under the License is distributed on an "AS IS" BASIS,
 *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 */

package io.github.basilapi.basil.sparql;

import java.util.Set;

public class SpecificationFactory {
	
	public static Specification create(String endpoint, String sparql) throws UnknownQueryTypeException {
		if(QueryType.isUpdate(sparql)) {
			return createUpdate(endpoint, sparql);
		} else {
			return createQuery(endpoint, sparql);
		}
	}
	
	private static Specification createQuery(String endpoint, String sparql) {
		VariablesCollector collector = new VariablesCollector();
		collector.collect(sparql);
		Set<String> vars = collector.getVariables();
		VariableParser parser;
		Specification spec = new Specification();
		spec.setEndpoint(endpoint);
		spec.setQuery(sparql);
		
		for (String var : vars) {
			parser = new VariableParser(var);
			if (parser.isParameter()) {
				spec.map(var, parser.getParameter());
			}
		}
		return spec;
	}
	
	private static Specification createUpdate(String endpoint, String sparql) {
		VariablesCollector collector = new VariablesCollector();
		collector.collect(sparql);
		Set<String> vars = collector.getVariables();
		VariableParser parser;
		Specification spec = new Specification();
		spec.setEndpoint(endpoint);
		spec.setUpdate(sparql);
		
		for (String var : vars) {
			parser = new VariableParser(var);
			if (parser.isParameter()) {
				spec.map(var, parser.getParameter());
			}
		}
		return spec;
	}
}
